+++ src/app/js/admin/progress/reducer.js
@@ -5,10 +5,12 @@ import { PUBLISH } from '../publish';
 export const UPDATE_PROGRESS = 'UPDATE_PROGRESS';
 export const ERROR_PROGRESS = 'ERROR_PROGRESS';
 export const LOAD_PROGRESS = 'LOAD_PROGRESS';
+export const FINISH_PROGRESS = 'FINISH_PROGRESS';
 
 export const updateProgress = createAction(UPDATE_PROGRESS);
 export const errorProgress = createAction(ERROR_PROGRESS);
 export const loadProgress = createAction(LOAD_PROGRESS);
+export const finishProgress = createAction(FINISH_PROGRESS);
 
 export const defaultState = {
     status: PENDING,
@@ -36,6 +38,7 @@ export default handleActions(
         }),
         [ERROR_PROGRESS]: (state, { payload: { error } }) => ({
             ...state,
+            status: PENDING,
             error,
         }),
     },
